refactor(#3991): extract isStaticDescWellFormed into descriptor-shape.ts
The LOC-regrowth ratchet (#3102/#3131) failed the PR: object-ops.ts grew
4789 -> 4826 (+37) because #3991's explanation of a subtle spec-routing decision
lived as a comment inside a 4,800-line god-file.

Taking the `loc-budget-allow:` grant would have been the wrong trade. The
predicate is a self-contained ToPropertyDescriptor (§6.2.5.6) classification with
no codegen state — it took a `ts.Expression` and returned a boolean while being a
closure nested inside `compileObjectDefineProperties` purely by history. Burying
it there is precisely how its central claim ("Object.defineProperty will handle
validation at runtime via its own path" — false) went unexamined long enough to
cost 347 test262 files. This mirrors #3984's extraction of
`maybeEmitVecLengthDefine` into array-length-define.ts, for the same reason.

The new module's header states the invariant the predicate exists to protect,
which the old inline version never did: **a `true` answer is a PROMISE that the
static expansion can fully model the descriptor**, not a "probably fine" — the
expansion has no fallback once it has claimed one, and its failure mode is a
silent wrong answer rather than a refusal.

Net effect on the god-file: 4789 -> 4737, a 52-line SHRINK rather than a 37-line
growth. `node scripts/check-loc-budget.mjs` passes with no allowance granted.

Pure code move — no behaviour change:
  - tsc --noEmit clean
  - all 29 measured test262 flips re-verified after the move (29/29), run solo
